Create 5_year_retun_stimulator
This project aims to create a portfolio stimulation to calculate the return for 40 popular stock and crypto assets. The program gives the user two choices to invest and will calculate the net profit/loss, and net portfolio value and will ask the user to choose the following two assets to invest. Once the user chooses to stop the stimulation it will output the net gain/loss, total portfolio value, and a line graph to visualize the portfolio growth over time. The return is based of the period from 2017 October 20 to 2022 October 20